refactor config & scripts
