--- src/task/methods/lb2/config.vsh.yaml
@@ -1,40 +1,0 @@
-__merge__: ../../api/comp_method.yaml
-
-functionality:
-  name: lb2
-
-  info:
-    label: lb2
-    summary: "This method is based on transformer and ensemble learning."
-    description: |
-      By treating it as a repression problem, this method utilize clusteirng sampling as pre-processing step and transformer as major model structure, to predict the response. The author also considers different cases and aggregate them into the best solution.
-    documentation_url: https://www.kaggle.com/competitions/open-problems-single-cell-perturbations/discussion/458738
-    repository_url: https://github.com/Eliorkalfon/single_cell_pb
-  resources:
-    - type: python_script
-      path: script.py
-    - path: models.py
-    - path: utils.py
-
-platforms:
-  - type: docker
-    image: ghcr.io/openproblems-bio/base_pytorch_nvidia:1.0.4
-    setup:
-      - type: python
-        packages:
-          - anndata
-          - torch==2.1.0
-          - torchvision==0.16.0
-          - torchaudio==2.1.0
-          - fastparquet
-          - pyarrow
-          - pandas~=2.0.3
-          - scikit-learn~=1.0.1
-          - tqdm~=4.66.1
-          - numpy~=1.23
-          - matplotlib~=3.5.0
-          - PyYAML~=6.0.1
-          - lion-pytorch
-  - type: nextflow
-    directives:
-      label: [ midtime, highmem, highcpu ]
