--- src/task/methods/lb2/requirements.txt
@@ -1,7 +1,0 @@
-pandas~=2.0.3
-scikit-learn~=1.0.1
-tqdm~=4.66.1
-numpy~=1.21.4
-matplotlib~=3.5.0
-PyYAML~=6.0.1
-lion-pytorch
