+++ src/task/methods/transformer_ensemble/config.vsh.yaml
@@ -1,0 +1,59 @@
+__merge__: ../../api/comp_method.yaml
+
+functionality:
+  name: transformer_ensemble
+  info:
+    label: Transformer Ensemble
+    rank: 2
+    summary: An ensemble of four transformer models, trained on diverse feature sets, with a cluster-based sampling strategy and robust validation for optimal performance.
+    description: |
+      This method employs an ensemble of four transformer models,
+      each with different weights and trained on slightly varying feature sets.
+      The feature engineering process involved one-hot encoding of categorical labels,
+      target encoding using mean and standard deviation, and enriching the feature set
+      with the standard deviation of target variables. Additionally, the dataset was
+      carefully examined to ensure data cleanliness. A sophisticated sampling strategy
+      based on K-Means clustering was employed to partition the data into training and
+      validation sets, ensuring a representative distribution. The model architecture
+      leveraged sparse and dense feature encoding, along with a transformer for effective
+      learning.
+    documentation_url: https://www.kaggle.com/competitions/open-problems-single-cell-perturbations/discussion/458738
+    repository_url: https://github.com/Eliorkalfon/single_cell_pb
+  arguments:
+    - name: --num_train_epochs
+      type: integer
+      default: 20000
+      description: "Number of training epochs."
+      info:
+        test_default: 10
+  resources:
+    - type: python_script
+      path: script.py
+    - path: models.py
+    - path: utils.py
+    - path: train.py
+    - path: predict.py
+    - path: seq.py
+
+platforms:
+  - type: docker
+    image: ghcr.io/openproblems-bio/base_pytorch_nvidia:1.0.4
+    setup:
+      - type: python
+        packages:
+          - anndata
+          - torch==2.1.0
+          - torchvision==0.16.0
+          - torchaudio==2.1.0
+          - fastparquet
+          - pyarrow
+          - pandas~=2.0.3
+          - scikit-learn~=1.0.1
+          - tqdm~=4.66.1
+          - numpy~=1.23
+          - matplotlib~=3.5.0
+          - PyYAML~=6.0.1
+          - lion-pytorch
+  - type: nextflow
+    directives:
+      label: [ midtime, highmem, highcpu ]
